Conflicts:
	src/plugins/platforms/xcb/qxcbconnection.h
	src/plugins/platforms/xcb/qxcbconnection_xi2.cpp
	src/plugins/sqldrivers/sqlite/qsql_sqlite.cpp
	src/plugins/styles/mac/qmacstyle_mac.mm
	src/widgets/widgets/qdockarealayout.cpp
	src/widgets/widgets/qmainwindow.cpp
	src/widgets/widgets/qmainwindowlayout.cpp
	src/widgets/widgets/qmainwindowlayout_p.h
	tests/auto/corelib/tools/qlocale/tst_qlocale.cpp
	tests/auto/other/macnativeevents/BLACKLIST
	tests/auto/widgets/widgets/qmenu/BLACKLIST

+++ src/widgets/dialogs/qfiledialog.cpp
@@ -3417,7 +3417,7 @@ void QFileDialogPrivate::_q_deleteCurrent()
 
     QModelIndexList list = qFileDialogUi->listView->selectionModel()->selectedRows();
     for (int i = list.count() - 1; i >= 0; --i) {
-        QModelIndex index = list.at(i);
+        QPersistentModelIndex index = list.at(i);
         if (index == qFileDialogUi->listView->rootIndex())
             continue;
 
@@ -3427,7 +3427,6 @@ void QFileDialogPrivate::_q_deleteCurrent()
 
         QString fileName = index.data(QFileSystemModel::FileNameRole).toString();
         QString filePath = index.data(QFileSystemModel::FilePathRole).toString();
-        bool isDir = model->isDir(index);
 
         QFile::Permissions p(index.parent().data(QFileSystemModel::FilePermissions).toInt());
 #if QT_CONFIG(messagebox)
@@ -3443,13 +3442,16 @@ void QFileDialogPrivate::_q_deleteCurrent()
                                       QMessageBox::Yes | QMessageBox::No, QMessageBox::No) == QMessageBox::No)
             return;
 
+        // the event loop has run, we have to validate if the index is valid because the model might have removed it.
+        if (!index.isValid())
+            return;
+
 #else
         if (!(p & QFile::WriteUser))
             return;
 #endif // QT_CONFIG(messagebox)
 
-        // the event loop has run, we can NOT reuse index because the model might have removed it.
-        if (isDir) {
+        if (model->isDir(index) && !model->fileInfo(index).isSymLink()) {
             if (!removeDirectory(filePath)) {
 #if QT_CONFIG(messagebox)
             QMessageBox::warning(q, q->windowTitle(),
